pyproject.toml
# project configuration, use poetry
[tool.poetry]
name = "lohrasb"
version = "4.0.0"
description = "This versatile tool streamlines hyperparameter optimization in machine learning workflows. It supports a wide range of search methods, from GridSearchCV and RandomizedSearchCV to advanced techniques like OptunaSearchCV, Ray Tune, and Scikit-Learn Tune. Designed to enhance model performance and efficiency, it's suitable for tasks of any scale."
authors = ["drhosseinjavedani <h.javedani@gmail.com>"]
homepage = "https://github.com/drhosseinjavedani/lohrasb"
license = "BSD-3-Clause license"

readme = "README.md"
keywords = [
    "Auto ML",
    "Pipeline",
    "Machine learning"
]


[tool.poetry.dependencies]
python = ">=3.8,<3.11"
numpy = "^1.20"
scipy = "^1.5"
optuna = "^3.2.0"
xgboost = "^1.6.1"
imblearn = "^0.0"
lightgbm = "^3.3.2"
nox = "^2022.1.7"
feature-engine = "^1.4.1"
category-encoders = "^2.5.0"
pandas = "^1.4"
xgbse = "^0.2.3"
pycox = "^0.2.3"
pyarrow = "^10.0.1"
scikit-learn = "^1.1.3"
interpret = "^0.3.0"
python-dotenv = "^0.21.1"
ray = {extras = ["tune"], version = "^2.2.0"}
tune-sklearn = "^0.4.5"
protobuf = "3.20"
hyperopt = "^0.2.7"
scikit-optimize = "^0.9.0"
bayesian-optimization = "^1.4.3"
catboost = "^1.2"
dill = "^0.3.6"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.6.0"
nox = "^2022.1.7"
bump2version = "^1.0.1"
isort = "^5.10.1"
pytest = "^7.1.2"
ipykernel = "^6.15.2"

[tool.poetry.group.dev.dependencies]
IProgress = "^0.4"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
